Move open telemetry scenario to its own file
